plocation

+++ stylehouse.pl
@@ -86,7 +86,7 @@ use Way;
 
 `cat /dev/null > proc/start`;
 `cat /dev/null > proc/list`;
-`rm proc/*.*`;
+`rm proc/*.*` if glob('proc/*.*');
 
 my ($name) = $Bin =~ m{/(\w+)$};
 say $Bin;
